chore(project): New build
